Explicitly specified major mode to use in file
Normally the Emacs config is in .emacs but our .emacs is actually a symlink
from $HOME/dotfiles/emacs

If ~/emacs is edited directly, emacs doesn't immediately detect that the
file is in Elisp, so emacs-elisp-mode isn't launched automatically. Of
course, you can manually start it with 'M-x emacs-elisp-mode', but to
start it automatically, you must define 'mode: emacs-lisp' at the top of
the file.

Also changed the formatting of the add'l Emacs package repos.