serve special dns address for all clients on 192.168.95.0
Don't know if this will work, but I am trying this to make D-I for
Ubuntu preseed install fully automated. Right now Ubuntu's D-I stops
to ask me for a nameserver address, but it is also ok to enter
nothing, which makes the install continue.

If dnsmasq can give a dns addy like 0.0.0.0 (black hole) to the
Ubuntu client then maybe D-I won't prompt me to enter a
nameserver address.