remove custom common lisp function for pkg ck & install
I used to use a custom CL function 'gojun-pkglist-installed-p'
in .emacs to check if a certain pkg was installed or not,
and install it if it wasn't installed.

This function depended on the now-deprecated "cl" module,
with "cl-lib" being its replacement. A google search revealed
a much simpler solution using native Elisp looping with
'dolist', so I removed the 'cl' dependency and just went with
an Elisp solution.