.emacs file for Windows
I setup Emacs for Windows on a Win11 VM today; this is my first time
installing an MS Windows-native Emacs; in the past I've used Emacs on
WSL (which is not really native Windows) and that worked OK. The biggest
difference is that the path to linters and Python interpreters uses
Windows absolute paths and the font settings are a bit different. I
manually installed 'monofur' TTF fonts and tried to get them to work
with Emacs and Windows, but it seems the monofur fonts I found on the
Internet are buggy; the TTF files render only in italics on Windows :(
so for the time being I'm using 'Consolas' as my default font.