.bashrc for root user, rm bad bashrc from Archlinux
- edited bashrc to be used in `/root/.bashrc`
- get rid of bashrc from Arch

The old .bashrc I used during my Archlinux days is an example of bad
practices that illustrate an ignorance of the differences between
bashrc and bash_profile. I have lots of env vars in this bashrc and
also invoke `ssh-agent` from bashrc which results in multiple agents
running everytime a an interactive non-login shell is invoked;

Fortunately today I run `ssh-agent` as a systemd user service, so
the messiness a la bashrc from Archlinux is no longer an issue.