Update dot emacs for MacOS
I haven't been keeping up with the latest changes in Emacs-land; apparently
the Fill Column Indicator, a visual indicator of line margins in the
buffer, was added to native Emacs in version 27, so it is no longer
necessary to use the 3rd-party package `fill-column-indicator`. I
therefore removed the package from the list of packages in `.emacs`
and the related settings.

I also added settings for the native fill-column-indicator to turn
it on in all buffers.

I removed the `sanityinc-solarized-light` theme and replaced it with
`solarized-theme`.

I also added `(server-mode 1)` so that Emacs will run as a server.
Then I will set shell env vars to use `emacsclient` instead of emacs.