edit 'org-after-todo-state-change-hook "CREATED:" str
Previously every time a new task is added and the task state is changed, a
"CREATED: <timestamp>" string was inserted into the task. The problem
with this is that the timestamp then showed up in my "org-agenda" for the
current day.

I therefore edited the inserted date to NOT use "org-time-stamp-format"
but instead use "strftime" format from C with square brackets around
the time so that org-mode will not treat this string as a timestamp.

Now this non-timestamp in my newly-created tasks no longer mistakenly
shows up in my org-agenda